@@ -2,30 +2,29 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.appdynamics.apmgame</groupId>
+    <groupId>com.appdynamics.apmgame.java</groupId>
     <artifactId>javanode</artifactId>
     <packaging>jar</packaging>
     <version>1.0-SNAPSHOT</version>
     <name>javanode</name>
     <properties>
-        <jettyVersion>[9.4.11.v20180605,)</jettyVersion>
+        <jettyVersion>11.0.0.beta3</jettyVersion>
     </properties>
     <dependencies>
-	    <dependency>
-    <groupId>javax.servlet</groupId>
-    <artifactId>javax.servlet-api</artifactId>
-    <version>3.0.1</version>
-    <scope>provided</scope>
-</dependency>
+        <dependency>
+            <groupId>jakarta.servlet</groupId>
+            <artifactId>jakarta.servlet-api</artifactId>
+            <version>5.0.0</version>
+        </dependency>
         <dependency>
             <groupId>mysql</groupId>
             <artifactId>mysql-connector-java</artifactId>
-            <version>5.1.47</version>
+            <version>5.1.49</version>
         </dependency>
         <dependency>
           <groupId>org.mongodb</groupId>
           <artifactId>mongo-java-driver</artifactId>
-          <version>3.10.2</version>
+          <version>3.12.7</version>
         </dependency>
         <dependency>
             <groupId>org.eclipse.jetty</groupId>
@@ -68,11 +67,11 @@
             <version>1.1.3</version>
         </dependency>
         <dependency>
-            <groupId>com.appdynamics</groupId>
-            <artifactId>appdynamics</artifactId>
-            <version>4.5</version>
-            <scope>system</scope>
-            <systemPath>${project.basedir}/javaagent.jar</systemPath>
+          <groupId>com.appdynamics</groupId>
+          <artifactId>appdynamics</artifactId>
+          <version>4.5</version>
+          <scope>system</scope>
+          <systemPath>${project.basedir}/javaagent.jar</systemPath>
         </dependency>
     </dependencies>
     <build>
@@ -132,10 +131,9 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.7.0</version>
+                <version>3.8.0</version>
                 <configuration>
-                    <source>1.7</source>
-                    <target>1.7</target>
+                    <release>11</release>
                 </configuration>
             </plugin>
         </plugins>
